Change splitting to try to split into as many parts as possible
Now if you give it 10 files and tell it to split it into 3 parts
it will not longer split it into 2 parts. It will split it into 3
parts and add the remaining files to the first part -> the first part
will have in this case one file more than the rest.
This and some other cases are now covered by tests.